Refactor settings navigation and components
- Updated all references from "/candidate/profile/settings" and "/company/settings" to "/settings" for a unified settings page.
- Removed the CandidateSettingsPage component and replaced it with a new SettingsShell component that dynamically loads the appropriate settings page based on user role.
- Created CompanySettingsPage and EmployeeSettingsPage components to handle company and employee-specific settings.
- Updated the dashboard and profile pages to link to the new settings page.
- Removed the old settings page file for candidates to streamline the codebase.
