--- src/pages/candidate/profile/settings.tsx
@@ -1,160 +1,0 @@
-"use client";
-import React, { useState, useCallback, useEffect } from "react";
-import { Box, Typography, Avatar } from "@mui/material";
-import { useTranslation } from "react-i18next";
-import CandidateWorkspaceLayout from "@/components/layout/candidate/CandidateWorkspaceLayout";
-import { NotificationsTab, PersonalInformationTab, ProfileVisibilityTab, SnackbarNotifications, LanguageTab, useProfileManagement, useUpdateCandidateVisibility } from "@/modules/settings/candidate";
-import PersonOutlined from "@mui/icons-material/PersonOutlined";
-import LanguageIcon from "@mui/icons-material/Language";
-import NotificationsOutlined from "@mui/icons-material/NotificationsOutlined";
-import VisibilityOutlined from "@mui/icons-material/VisibilityOutlined";
-
-import { TEAL as T, TEAL_BG as TBG, TEAL_BORDER as TBRD } from "@/modules/settings/shared";
-
-const TL   = "#14B8A6";
-const NAVY = "#0D1B2A";
-
-const TAB_IDS = [
-  { id: "personal",      key: "personal",      icon: PersonOutlined },
-  { id: "language",      key: "language",       icon: LanguageIcon },
-  { id: "notifications", key: "notifications",  icon: NotificationsOutlined },
-  { id: "visibility",    key: "visibility",     icon: VisibilityOutlined },
-];
-
-const CandidateSettingsPage: React.FC = () => {
-  const { t } = useTranslation("dashboard");
-
-  const TABS = TAB_IDS.map(tab => ({
-    ...tab,
-    label: t(`candidate_settings.tabs.${tab.key}`),
-  }));
-  const {
-    activeTab, isEditing, profile, loading, error,
-    uploadingImage, saveSuccess, userId,
-    companyMembership, isPublicProfile,
-    control, formErrors,
-    setActiveTab, setIsEditing,
-    handleInputChange, handleImageUpload,
-    handleSaveProfile, handleSaveLanguage, handleCancel, handleDismissError, handleDismissSuccess,
-  } = useProfileManagement();
-
-  const hasMembership = !!companyMembership?._id;
-
-  const updateVisibilityMutation = useUpdateCandidateVisibility();
-  const [localIsPublic, setLocalIsPublic] = useState(isPublicProfile);
-
-  useEffect(() => { setLocalIsPublic(isPublicProfile); }, [isPublicProfile]);
-
-  const handleToggleVisibility = useCallback(async (newVisibility: boolean) => {
-    setLocalIsPublic(newVisibility);
-    await updateVisibilityMutation.mutateAsync(newVisibility);
-  }, [updateVisibilityMutation]);
-
-  const displayName = profile.firstName
-    ? `${profile.firstName}${profile.lastName ? ` ${profile.lastName}` : ""}`
-    : profile.username || "Candidate";
-  const initial   = displayName[0]?.toUpperCase() || "C";
-  const avatarUrl = profile.avatar || undefined;
-
-  return (
-    <CandidateWorkspaceLayout breadcrumb={t("candidate.nav.settings")}>
-      <Box sx={{ display: "grid", gridTemplateColumns: { xs: "1fr", md: "240px 1fr" }, gap: 2.5, alignItems: "start" }}>
-
-        {/* LEFT: Profile + Nav */}
-        <Box sx={{ display: { xs: "none", md: "flex" }, flexDirection: "column", gap: 2, position: "sticky", top: 16, maxHeight: "calc(100vh - 96px)", overflowY: "auto" }} className="custom-scrollbar">
-
-            {/* Profile card */}
-            <Box sx={{ bgcolor: "#fff", borderRadius: "16px", border: "1px solid #E5E7EB", overflow: "hidden", boxShadow: "0 1px 4px rgba(0,0,0,0.04)" }}>
-              <Box sx={{ height: 56, background: `linear-gradient(135deg, ${NAVY} 0%, ${T} 100%)`, position: "relative" }}>
-                <Box sx={{ position: "absolute", top: "50%", right: 16, transform: "translateY(-50%)", width: 32, height: 32, borderRadius: "50%", bgcolor: `${TL}30`, border: `1px solid ${TL}40` }} />
-              </Box>
-              <Box sx={{ px: 2, pb: 2 }}>
-                <Box sx={{ mt: -3, mb: 1 }}>
-                  <Avatar src={avatarUrl} sx={{ width: 52, height: 52, bgcolor: T, fontSize: "1.2rem", fontWeight: 700, border: "2.5px solid #fff", boxShadow: "0 2px 8px rgba(0,0,0,0.12)" }}>
-                    {initial}
-                  </Avatar>
-                </Box>
-                <Typography sx={{ fontWeight: 800, fontSize: "0.95rem", color: NAVY, lineHeight: 1.2 }}>{displayName}</Typography>
-                {profile.email && <Typography sx={{ fontSize: "0.72rem", color: "#9CA3AF", mt: 0.25 }}>{profile.email}</Typography>}
-              </Box>
-            </Box>
-
-            {/* Tab nav */}
-            <Box sx={{ bgcolor: "#fff", borderRadius: "16px", border: "1px solid #E5E7EB", p: 1.25, boxShadow: "0 1px 4px rgba(0,0,0,0.04)" }}>
-              <Typography sx={{ fontSize: "0.65rem", fontWeight: 700, color: "#94A3B8", textTransform: "uppercase", letterSpacing: "0.06em", px: 1, pb: 1 }}>{t("candidate_settings.tabs.settings_label")}</Typography>
-              <Box sx={{ display: "flex", flexDirection: "column", gap: 0.5 }}>
-                {TABS.map(({ id, label, icon: Icon }) => {
-                  const active = activeTab === id;
-                  return (
-                    <Box
-                      key={id}
-                      onClick={() => setActiveTab(id)}
-                      sx={{
-                        display: "flex", alignItems: "center", gap: 1.25,
-                        px: 1.5, py: 1, borderRadius: "10px", cursor: "pointer",
-                        bgcolor: active ? TBG : "transparent",
-                        border: `1px solid ${active ? TBRD : "transparent"}`,
-                        transition: "all 0.15s",
-                        "&:hover": { bgcolor: active ? TBG : "#F8FAFC", borderColor: active ? TBRD : "#E5E7EB" },
-                      }}
-                    >
-                      <Icon sx={{ fontSize: 16, color: active ? T : "#6B7280" }} />
-                      <Typography sx={{ fontSize: "0.82rem", fontWeight: active ? 700 : 500, color: active ? T : "#374151" }}>{label}</Typography>
-                      {active && <Box sx={{ ml: "auto", width: 6, height: 6, borderRadius: "50%", bgcolor: T }} />}
-                    </Box>
-                  );
-                })}
-              </Box>
-            </Box>
-          </Box>
-
-          {/* RIGHT: Tab content */}
-          <Box>
-            {activeTab === "personal" && (
-              <PersonalInformationTab
-                profile={profile}
-                control={control}
-                formErrors={formErrors}
-                isEditing={isEditing}
-                loading={loading}
-                saveSuccess={saveSuccess}
-                error={error || null}
-                uploadingImage={uploadingImage}
-                onImageUpload={handleImageUpload}
-                onSave={handleSaveProfile}
-                onCancel={handleCancel}
-                onEditToggle={() => setIsEditing(!isEditing)}
-              />
-            )}
-            {activeTab === "language" && (
-              <LanguageTab
-                onInputChange={(key, value) => handleInputChange(key as keyof typeof profile, value)}
-                onSaveLanguage={handleSaveLanguage}
-                showGenerateLanguage={false}
-                centerInterfaceVertically={true}
-              />
-            )}
-            {activeTab === "notifications" && <NotificationsTab />}
-            {activeTab === "visibility" && (
-              <ProfileVisibilityTab
-                userId={userId}
-                effectiveIsPublicProfile={localIsPublic}
-                onToggleVisibility={handleToggleVisibility}
-                hasMembership={hasMembership}
-              />
-            )}
-          </Box>
-
-        </Box>
-
-        <SnackbarNotifications
-          error={error || null}
-          saveSuccess={saveSuccess}
-          onDismissError={handleDismissError}
-          onDismissSuccess={handleDismissSuccess}
-        />
-      </CandidateWorkspaceLayout>
-    );
-  };
-
-  export default CandidateSettingsPage;
