+++ src/pages/settings/index.tsx
@@ -1,0 +1,5 @@
+import SettingsShell from "@/modules/settings/shared/components/SettingsShell";
+
+export default function SettingsPage() {
+  return <SettingsShell />;
+}
